Implicitly initialize the migrations directory
Feature request from the CLI side.
