+++ libs/test-setup/src/lib.rs
@@ -1,3 +1,5 @@
+#![deny(unsafe_code, rust_2018_idioms)]
+
 //! This crate contains constants and utilities that are useful for writing tests across the
 //! engines.
